+++ lib/mongoid/contextual/eager.rb
@@ -10,6 +10,8 @@ module Mongoid
 
       # Eager load the inclusions for the provided documents.
       #
+      # @api private
+      #
       # @example Eager load the inclusions.
       #   context.eager_load(docs)
       #
@@ -19,13 +21,25 @@ module Mongoid
       #
       # @since 3.0.0
       def eager_load(docs)
-        # @todo: #2638: We've rejected all the inclusions after calling #first
-        # or #last, so when asking for the entire criteria after none of these
-        # remain to get executed.
-        criteria.inclusions.reject! do |metadata|
-          metadata.eager_load(eager_loaded_ids(docs, metadata)) if !docs.empty?
-        end
+        load_inclusions(docs)
         self.eager_loaded = true
+      end
+
+      # Eager load the inclusions for the provided document.
+      #
+      # @api private
+      #
+      # @example Eager load the inclusions.
+      #   context.eager_load(doc)
+      #
+      # @param [ Document ] doc The doc returning from the db.
+      #
+      # @return [ true ] Always true.
+      #
+      # @since 3.0.16
+      def eager_load_one(doc)
+        load_inclusions([ doc ])
+        inclusions_loaded[doc.id] = true
       end
 
       # Get the ids that to be used to eager load documents.
@@ -51,18 +65,77 @@ module Mongoid
 
       # Is this context able to be eager loaded?
       #
+      # @api private
+      #
       # @example Is the context eager loadable?
       #   context.eager_loadable?
+      #
+      # @example Is the single document eager loadable?
+      #   context.eager_loadable?(document)
+      #
+      # @param [ Document ] document The single document to load for.
       #
       # @return [ true, false ] If the context is able to be eager loaded.
       #
       # @since 3.0.0
-      def eager_loadable?
-        !eager_loaded && !criteria.inclusions.empty?
+      def eager_loadable?(document = nil)
+        return false if criteria.inclusions.empty?
+        document ? !inclusions_loaded?(document) : !eager_loaded
+      end
+
+      # Has a hash of individual documents that have had their relations reager
+      # loaded.
+      #
+      # @api private
+      #
+      # @example Get the documents with relations eager loaded.
+      #   context.inclusions_loaded
+      #
+      # @return [ Hash ] The documents that have had eager loaded inclusions.
+      #
+      # @since 3.0.16
+      def inclusions_loaded
+        @inclusions_loaded ||= {}
+      end
+
+      # Has the document had its inclusions loaded?
+      #
+      # @api private
+      #
+      # @example Has the document had its inclusions loaded?
+      #   context.inclusions_loaded?(document)
+      #
+      # @param [ Document ] document The document to check.
+      #
+      # @return [ true, false ] If the document had it's inclusions loaded.
+      #
+      # @since 3.0.16
+      def inclusions_loaded?(document)
+        inclusions_loaded.has_key?(document.id)
+      end
+
+      # Eager load the inclusions for the provided documents.
+      #
+      # @api private
+      #
+      # @example Eager load the inclusions.
+      #   context.load_inclusions(docs)
+      #
+      # @param [ Array<Document> ] docs The docs returning from the db.
+      #
+      # @return [ true ] Always true.
+      #
+      # @since 3.0.16
+      def load_inclusions(docs)
+        criteria.inclusions.each do |metadata|
+          metadata.eager_load(eager_loaded_ids(docs, metadata)) if !docs.empty?
+        end
       end
 
       # If the provided document exists, eager load its dependencies or return
       # nil.
+      #
+      # @api private
       #
       # @example Eager load if the document is not nil.
       #   context.with_eager_loading(document)
@@ -76,7 +149,7 @@ module Mongoid
         selecting do
           return nil unless document
           doc = Factory.from_db(klass, document, criteria.object_id)
-          eager_load([ doc ]) if eager_loadable?
+          eager_load_one(doc) if eager_loadable?(doc)
           doc
         end
       end