+++ spec/mongoid/criteria_spec.rb
@@ -2543,7 +2543,7 @@ describe Mongoid::Criteria do
         end
 
         before do
-          context.should_receive(:eager_load).with([ post_one ]).once.and_call_original
+          context.should_receive(:eager_load_one).with(post_one).once.and_call_original
         end
 
         let!(:document) do
@@ -2574,7 +2574,7 @@ describe Mongoid::Criteria do
         end
 
         before do
-          context.should_receive(:eager_load).with([ post_two ]).once.and_call_original
+          context.should_receive(:eager_load_one).with(post_two).once.and_call_original
         end
 
         let!(:document) do
@@ -2658,7 +2658,7 @@ describe Mongoid::Criteria do
           end
 
           before do
-            new_context.should_receive(:eager_load).with([ person ]).once.and_call_original
+            new_context.should_receive(:eager_load_one).with(person).once.and_call_original
           end
 
           let!(:from_db) do
@@ -2694,7 +2694,7 @@ describe Mongoid::Criteria do
         end
 
         before do
-          context.should_receive(:eager_load).with([ person ]).once.and_call_original
+          context.should_receive(:eager_load_one).with(person).once.and_call_original
         end
 
         let!(:from_db) do
@@ -2729,7 +2729,7 @@ describe Mongoid::Criteria do
         end
 
         before do
-          context.should_receive(:eager_load).with([ person ]).once.and_call_original
+          context.should_receive(:eager_load_one).with(person).once.and_call_original
         end
 
         let!(:from_db) do
@@ -2861,7 +2861,7 @@ describe Mongoid::Criteria do
         end
 
         before do
-          context.should_receive(:eager_load).with([ person ]).once.and_call_original
+          context.should_receive(:eager_load_one).with(person).once.and_call_original
         end
 
         let!(:from_db) do
@@ -2900,7 +2900,7 @@ describe Mongoid::Criteria do
         end
 
         before do
-          context.should_receive(:eager_load).with([ person ]).once.and_call_original
+          context.should_receive(:eager_load_one).with(person).once.and_call_original
         end
 
         let!(:from_db) do
@@ -3036,7 +3036,7 @@ describe Mongoid::Criteria do
         end
 
         before do
-          context.should_receive(:eager_load).with([ person ]).once.and_call_original
+          context.should_receive(:eager_load_one).with(person).once.and_call_original
         end
 
         let!(:from_db) do
@@ -3053,6 +3053,21 @@ describe Mongoid::Criteria do
 
         it "inserts the second document into the identity map" do
           Mongoid::IdentityMap[Post.collection_name][post_two.id].should eq(post_two)
+        end
+
+        context "when subsequently getting all documents" do
+
+          before do
+            context.should_receive(:eager_load).with([ person ]).once.and_call_original
+          end
+
+          let!(:documents) do
+            criteria.entries
+          end
+
+          it "returns the correct documents" do
+            documents.should eq([ person ])
+          end
         end
       end
 
@@ -3071,7 +3086,7 @@ describe Mongoid::Criteria do
         end
 
         before do
-          context.should_receive(:eager_load).with([ person ]).once.and_call_original
+          context.should_receive(:eager_load_one).with(person).once.and_call_original
         end
 
         let!(:from_db) do
@@ -3088,6 +3103,21 @@ describe Mongoid::Criteria do
 
         it "inserts the second document into the identity map" do
           Mongoid::IdentityMap[Post.collection_name][post_two.id].should eq(post_two)
+        end
+
+        context "when subsequently getting all documents" do
+
+          before do
+            context.should_receive(:eager_load).with([ person ]).once.and_call_original
+          end
+
+          let!(:documents) do
+            criteria.entries
+          end
+
+          it "returns the correct documents" do
+            documents.should eq([ person ])
+          end
         end
       end